Refactoring install.sh and add Alarm Clock in main Menu desctop

+++ install_alarm.sh
@@ -1,15 +1,41 @@
 #! /usr/bin/sh
-mkdir build
-cd ./build || exit
-rm -R *
-cd ../src || exit
-javac -verbose -d ../build Main.java
-cp main/resources/electra.ttf ~/.fonts
-cd ..
-cp alarmClock.sh ./build/alarmClock.sh
-command1='cd '`pwd`/build'; '
-command2='java Main $@'
-fullCommand='alias alarmClock="'$command1$command2'"'
-echo $fullCommand >> ~/.bashrc
-source ~/.bashrc
 
+if javac --version
+    
+then
+    echo 'Found jdk. Start compiling classes in ./src'
+    mkdir build
+    cd ./build || exit
+    rm -R *
+    cd ../src || exit
+    javac -verbose -d ../build Main.java
+    cp main/resources/electra.ttf ~/.fonts
+    cd ..
+    command1='java -classpath '`pwd`/build' Main $@'
+    fullCommand='alias alarmClock="'$command1'"'
+    echo $fullCommand
+    echo $fullCommand >> ~/.bashrc
+    . ~/.bashrc
+
+
+    desktopFile=~/.local/share/applications/alarmClock.desktop
+    if find ~/.local/share/applications/ -name alarmClock.desktop
+        then 
+            echo $desktopFile" Found. Removing it."
+            rm $desktopFile
+    fi
+    echo $desktopFile" Create"
+    touch $desktopFile
+    echo '[Desktop Entry]' >> $desktopFile
+    echo 'Type=Application' >> $desktopFile
+    echo 'Encoding=UTF-8' >> $desktopFile
+    echo 'Name=Simple Alarm Clock' >> $desktopFile
+    echo 'Comment=Alarm clock, Timer and Stop Watch' >> $desktopFile
+    echo 'Exec=java -classpath '`pwd`'/build Main' >> $desktopFile
+    echo 'Terminal=false' >> $desktopFile
+
+
+else
+    echo 'Not found jdk. Please install and try again!'
+    exit 1
+fi
